Open Boosters
Started fleshing out the Open Boosters page, which for now just displays purchased product and lets you select which one to interact with (interoping correctly with the Collection Manager).
Settings now both deletes the collection and refreshes the definitions manually.
Generalized the definition loading so any page could access it.